use nodo for running shiki
Cache the highlighter instance in a script and defined a highlight
function to generate the html
